AI_Employee_Vault/requirements.txt
# AI Employee - Silver Tier Requirements

# Core (Bronze Tier)
watchdog>=3.0.0

# Gmail API
google-api-python-client>=2.100.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=1.0.0

# WhatsApp & LinkedIn (Browser Automation)
playwright>=1.40.0

# Utilities
python-dotenv>=1.0.0
